refactor: LoginMemberService의 의존성 제거
