+++ src/main/java/nextstep/member/application/MemberService.java
@@ -1,13 +1,15 @@
 package nextstep.member.application;
 
+import nextstep.auth.application.LoginMemberService;
 import nextstep.member.application.dto.MemberRequest;
 import nextstep.member.application.dto.MemberResponse;
+import nextstep.member.domain.LoginMember;
 import nextstep.member.domain.Member;
 import nextstep.member.domain.MemberRepository;
 import org.springframework.stereotype.Service;
 
 @Service
-public class MemberService {
+public class MemberService implements LoginMemberService {
     private MemberRepository memberRepository;
 
     public MemberService(MemberRepository memberRepository) {
@@ -46,4 +48,10 @@ public class MemberService {
     public void deleteMember(String email) {
         memberRepository.deleteByEmail(email);
     }
+
+    public LoginMember findByUsername(String email) {
+        Member member = memberRepository.findByEmail(email).orElseThrow(RuntimeException::new);
+        return LoginMember.of(member);
+    }
+
 }
